+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## ???
-- Fix synchronization bug from project copy to working directory
+- Log to output channel
+- Fix synchronization bug (from project copy to working directory)
 
 ## 1.0.1
 Initial release